Laba7

+++ app/Http/Controllers/ArticleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Article;
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -26,6 +27,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', [self::class]);
         return view('articles.create');
     }
 
@@ -71,6 +73,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('create', [self::class]);
         return view('articles.edit', ['article' => $article]);
     }
 
@@ -104,6 +107,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('create', [self::class]);
         Comment::where('article_id', $article->id)->delete();
         $article->delete();
         return redirect('/');
